+++ examples/composition.rs
@@ -54,7 +54,9 @@ fn main() {
     let codecs = ImageCodec::built_in_codecs();
     img.write_to_file(
         "bl-getting-started-5.bmp",
-        codecs.find_codec_by_name("BMP").unwrap(),
+        ImageCodec::built_in_codecs()
+            .find_codec_by_name("BMP")
+            .unwrap(),
     )
     .expect("Writing to file failed");
 }