+++ examples/path.rs
@@ -27,10 +27,11 @@ fn main() {
     render(ctx).expect("Rendering to context failed");
 
     // Let's use some built-in codecs provided by Blend2D.
-    let codecs = ImageCodec::built_in_codecs();
     img.write_to_file(
         "bl-getting-started-1.bmp",
-        codecs.find_codec_by_name("BMP").unwrap(),
+        ImageCodec::built_in_codecs()
+            .find_codec_by_name("BMP")
+            .unwrap(),
     )
     .expect("Writing to file failed");
 }