+++ examples/rust_bl_logo.rs
@@ -99,10 +99,11 @@ fn main() {
     };
     render(ctx).expect("Rendering to context failed");
 
-    let codecs = ImageCodec::built_in_codecs();
     img.write_to_file(
         "rust_bl_logo.bmp",
-        codecs.find_codec_by_name("BMP").unwrap(),
+        ImageCodec::built_in_codecs()
+            .find_codec_by_name("BMP")
+            .unwrap(),
     )
     .expect("Writing to file failed");
 }